refactor: 모든 AcceptanceTest 리팩토링
- 모든 AcceptanceTest의 시나리오 포맷 통일
- MemberAcceptanceTest 패키지 변경
- RaceAcceptanceTest 시나리오에 Mission 생성 내용 추가
